Update supported_llm_providers.md
I replaced OpenAICallParams model for a openai model name (this string did not matter, but if someone changes provider, it will be a missconfiguration).

Also, i added other arguments to azure_client_wrapper that were necessary for my deployment to work. If set only the arguments provided on the example, it would return resource not found.